run concurrent goroutines (max `maxGoroutines` count at a time)
since we are limiting the number of concurrent goroutines, we have to
either drop urls, store them in a queue. Here I introduce one more flag,
which could be used to set size of the urls queue (100 by default). If
the user reaches that limit (e.g. by processing huge sites), new urls
will be dropped.